Split up LidarObstacle and LidarDecision
- split them into seperate header files, as well as seperate tests
